requirements.txt
Flask[async]==3.0.2
python-telegram-bot[webhooks]==20.8
httpx==0.26.0
python-dotenv==1.0.1
pytz==2024.1
gspread==5.12.4
oauth2client==4.1.3
nest_asyncio==1.6.0
holidays==0.92